Remove usage of TGTDIR in compiler files
Variable TGTDIR is defined in the build makefile but is used in the compiler
makefiles which are included from the main makefile, thus it cannot have been
defined when it was used in the compiler files. Problem detected on switch to
make-4.0.

The usage of TGTDIR in the compiler files was to modify the link flags. The
modification to the link flags is generic and can be/is better solved in the
compile and link target makefile.