Move commands for rpm building to build makefile
Rpm building is relevant in meta mode as well but calling the build stage from
meta stage when command rpmbuild is missing will now result in a relevant
error message rather then a "no such target" as was the previous case.

Condition if an rpm-target is called is used for minimizing make code, expcept
for clean targets for intermediate rpm-files.