Fix broken implicit internal include paths
Include paths to internal libraries (libs which are built within the project)
are automatically derived by looking for the 'include' directory for each
library a target is linked against. This must be performed at the end as
pre-processing stage when all targets have been found, it is only by then Bob
can know about all include directories. Because of this the variable used for
holding this must be a delayed variable. (This variable had ended up in an
eval-statement and thus become empty).