Remove recursive source file fetching macro
Recursive fetching of sources via macro should not be needed, better to use
glob-expressions. For implicit sources it could be useful to traverse the
full tree, but in a case where so many sources files are required that extra
directories are needed there should also be a way to more carefully select
what files are built. Anyhow this removes a macro which has never really been
used, i.e. making things less complicated.