Revert "Simplify dirprefix macro"
This reverts commit 494fcb8261a4c6c06d0463823ea47c56c49e6eb0.

Intriguing parts break when uic rules are setup without exact matching paths.
Macro setup_uic_depend_rules_srcs which depends all source files to the result
of the uic-process (i.e. the header file) did not match exactly what is setup
in macro setup_forms_rules_qt4 which setup the rules for creating the UI
header files. Resulting in that the source files did not depend in the UI
header and therefore it was not created. I.e. no one was needing the UI header
file.